管理画面で招待、所属削除の実装・所属削除のロジック修正

+++ app/controllers/companies_controller.rb
@@ -23,11 +23,9 @@ class CompaniesController < ApplicationController
   end
 
   def destroy_member
-    if @company.admin_user?(current_user)
-      employee = company.employees.find_by(user_id: params[:user_id])
-      employee.destroy
-    end
-    redirect_to company_ideas_path(company)
+    employee = @company.remove_employee(current_user, params[:user_id])
+    employee.destroy if employee
+    redirect_to company_ideas_path(@company)
   end
 
   private
